Move random_uuid_str to vendor_parser
