+++ src/allotropy/parsers/appbio_quantstudio/appbio_quantstudio_structure.py
@@ -22,7 +22,6 @@ from allotropy.allotrope.pandas_util import read_csv
 from allotropy.parsers.appbio_quantstudio.calculated_document import CalculatedDocument
 from allotropy.parsers.appbio_quantstudio.referenceable import Referenceable
 from allotropy.parsers.lines_reader import LinesReader
-from allotropy.parsers.utils.uuids import random_uuid_str
 from allotropy.parsers.utils.values import (
     assert_not_empty_df,
     assert_not_none,
@@ -37,6 +36,7 @@ from allotropy.parsers.utils.values import (
     try_str_from_series_or_default,
     try_str_from_series_or_none,
 )
+from allotropy.parsers.vendor_parser import random_uuid_str
 
 
 @dataclass(frozen=True)
