+++ src/allotropy/parsers/chemometec_nucleoview/nucleoview_parser.py
@@ -36,8 +36,7 @@ from allotropy.parsers.chemometec_nucleoview.constants import (
     NUCLEOCOUNTER_SOFTWARE_NAME,
 )
 from allotropy.parsers.chemometec_nucleoview.nucleoview_reader import NucleoviewReader
-from allotropy.parsers.utils.uuids import random_uuid_str
-from allotropy.parsers.vendor_parser import VendorParser
+from allotropy.parsers.vendor_parser import random_uuid_str, VendorParser
 
 _PROPERTY_LOOKUP = {
     "Dead (cells/ml)": TQuantityValueMillionCellsPerMilliliter,
