+++ src/allotropy/parsers/example_weyland_yutani/example_weyland_yutani_parser.py
@@ -22,8 +22,7 @@ from allotropy.parsers.example_weyland_yutani.example_weyland_yutani_structure i
     Data,
 )
 from allotropy.parsers.lines_reader import CsvReader, read_to_lines
-from allotropy.parsers.utils.uuids import random_uuid_str
-from allotropy.parsers.vendor_parser import VendorParser
+from allotropy.parsers.vendor_parser import random_uuid_str, VendorParser
 
 
 class ExampleWeylandYutaniParser(VendorParser):
