+++ src/allotropy/parsers/perkin_elmer_envision/perkin_elmer_envision_structure.py
@@ -29,7 +29,6 @@ from allotropy.allotrope.models.plate_reader_benchling_2023_09_plate_reader impo
 from allotropy.allotrope.models.shared.components.plate_reader import SampleRoleType
 from allotropy.exceptions import AllotropeConversionError
 from allotropy.parsers.lines_reader import CsvReader
-from allotropy.parsers.utils.uuids import random_uuid_str
 from allotropy.parsers.utils.values import (
     assert_not_none,
     try_float_from_series,
@@ -37,6 +36,7 @@ from allotropy.parsers.utils.values import (
     try_str_from_series,
     try_str_from_series_or_none,
 )
+from allotropy.parsers.vendor_parser import random_uuid_str
 
 
 def df_to_series(df: pd.DataFrame) -> pd.Series[Any]:
