--- src/allotropy/parsers/utils/uuids.py
@@ -1,5 +1,0 @@
-import uuid
-
-
-def random_uuid_str() -> str:
-    return str(uuid.uuid4())
